+++ src/Pim/Bundle/CatalogBundle/Datagrid/ProductDatagridManager.php
@@ -93,6 +93,11 @@ class ProductDatagridManager extends FlexibleDatagridManager
             'field'  => FieldDescriptionInterface::TYPE_OPTIONS,
             'filter' => FilterInterface::TYPE_CURRENCY
         );
+
+        self::$typeMatches['metric'] = array(
+            'field'  => FieldDescriptionInterface::TYPE_TEXT,
+            'filter' => FilterInterface::TYPE_METRIC
+        );
     }
 
     /**
@@ -277,6 +282,9 @@ class ProductDatagridManager extends FlexibleDatagridManager
         $backendType = $attribute->getBackendType();
         if ($backendType !== AbstractAttributeType::BACKEND_TYPE_OPTION
             && $result['type'] === FieldDescriptionInterface::TYPE_OPTIONS) {
+            $result['sortable'] = false;
+        } elseif ($backendType === AbstractAttributeType::BACKEND_TYPE_METRIC) {
+            $result['field_options']['family'] = $attribute->getMetricFamily();
             $result['sortable'] = false;
         }
 
@@ -596,6 +604,7 @@ class ProductDatagridManager extends FlexibleDatagridManager
             ->leftJoin($rootAlias.'.values', 'values')
             ->leftJoin('values.options', 'valueOptions')
             ->leftJoin('values.prices', 'valuePrices')
+            ->leftJoin('values.metric', 'valueMetrics')
             ->leftJoin($rootAlias .'.categories', 'category');
 
         $familyExpr = "(CASE WHEN ft.label IS NULL THEN productFamily.code ELSE ft.label END)";
@@ -604,6 +613,7 @@ class ProductDatagridManager extends FlexibleDatagridManager
             ->addSelect('values')
             ->addSelect('valuePrices')
             ->addSelect('valueOptions')
+            ->addSelect('valueMetrics')
             ->addSelect('category')
             ->addSelect('pGroup');